Renamed to SwiftWeather, and removed abbreviations from names.
I intend to change the API significantly, such that keeping the "SwiftyWeatherKit" name might lead to confusion vs the other forks and the original.

I also find the old name to be a bit of a mouthful.  I'm not entirely thrilled about "SwiftWeather", in fact - "Weather" would be more concise - but there's the SEO angle to consider.  Sigh.