Added proper formatting support to WeatherSensor.
It follows the design cues of various Foundation data types (e.g. Measurement).

It contains at least one sharp edge, though - it's seemingly very difficult, if not impossible, to actually implement Codable support for the formatter, which is required by the FormatStyle protocol.  So for now it just throws exceptions if any encoding or decoding is attempted.

I would like Codable support - you can imagine it'd be very convenient for working with user preferences re. presentation format - but I don't see a practical path to it.  For now.