Removed the dependency on Alamofire and SwiftyJSON.
I originally intended to only remove SwiftyJSON at this point, since it seemed trivial in the way it was being used and easily replaced with Foundation's standard JSON decoding.  Unfortunately a seemingly minor code change, in the process of doing that, resulted in the Swift compiler throwing completely useless and impenetrable error messages.  Long story short, I _thought_ it was something specific to the use of Alamofire, so I went ahead and removed that anyway.  Turns out it was actually the compiler trying - and failing spectacularly - to simply say that a closure was throwing when it wasn't supposed to.  Sigh. 🙄

There ended up being a bunch of entangled changes as part of that, all kind of cascading from each other in a codependent way.  So this patch is a bit messy. 😕

A few notable callouts:

  • The code now tries optimistically decoding the response as the expected format, rather than explicitly checking for an error response first.  It only looks for an error response - and even then only as an aid to selecting the best error code - if there's a failure.  This seems better, as it doesn't waste time in the common case - where there are no problems - in trying & failing to decode the wrong schema.

  • Moved the aforementioned error-message extraction into a new helper static method on AmbientWeatherError, from(apiResponse:else:).  It will try to determine the precise error reason from the response, but if it can't (e.g. it's missing or in an unexpected format) it has fallbacks.

  • Now actually check the HTTP response codes and behave appropriately if they're non-200s.

  • Made deviceEndPoint() and dataEndPoint():
	• Return URLs instead of Strings.
	• Throw more precise error messages.
	• Use guard syntax.

  • Made WeatherServiceStatus' Error case have an associated value of type Error, so as to convey some actual information about what happened.