Replaced 'reportingDevices' with 'devices'.
This is both a name change and a type change - now it's an optional map from device IDs to device [instances], instead of a non-optional list of such mappings (which in practice saw only one device per dictionary - very weird data structure).

This does mean that it's genuinely not possible to properly support multiple devices with conflicting IDs, but this is intrinsically the case pretty much by definition - the ID is _supposed_ to be a unique identifier, and APIs like AmbientWeather's use it to identify the weather station that is the subject of API calls, so there really is nothing sane or graceful that can be done if the API reports conflicting devices.