Removed the 'devices' property from WeatherPlatform and instead put that mapping into the Reporting status case.
This way WeatherPlatforms are essentially stateless (as far as users are concerned, at least).  It's also more elegant for the time period before the first successful retrieval of the station list - the previous 'devices' property had to be optional just to express this brief period, which made using it a bit of a pain anyway.