Replaced WeatherPlatform's "setupService" method with a simple "devices" property, and added "platform" to WeatherDevice.
This also allowed WeatherServiceStatus to be eliminated entirely.

This is the first phase of refactoring the API to be much more idiomatic Swift (e.g. use of async & plain throws) and to lay things out more logically - shortly the means for getting weather reports will move from WeatherPlatform to WeatherDevice where it makes much more sense.