Revamped the API for fetching reports.
It's now based out of WeatherDevices, not WeatherPlatforms, since the reports are device-specific.

The methods (and a convenience property for simply getting the latest report) are all async and throwing per idiomatic Swift.

As part of this I added the ability to specify the end date of the report search range.

I also implemented the AmbientWeatherDevice's version of the new API such that it automatically overcomes the 288-report-per-API-call limit imposed by AmbientWeather, by breaking the task up into multiple 288-or-smaller chunks.

One note:  I utilised AsyncThrowingStream as the means to return an AsyncSequence, but I'm not sure that's the best way.  It feels non-native to be using a continuation, and requires a bit of boilerplate.  But the only other option appears to be to implement a custom struct conforming to AsyncStream, which seems like quite a big jump in complexity.  Plus the API is ill-defined re. iterators and consumption. 😕