Added a documentation comment about the possibility that returned AsyncThrowingStreams may run ahead of consumption.
I don't actually want that to be the case, at least not unboundedly - but AsyncThrowingStream has no viable way to prevent this; I'd have to implement a custom AsyncSequence.  Which I might do in future… I just don't want to get into it right now as it's not top priority.