Added the static function 'fakeDeviceForTesting' to AmbientWeatherDevice.
This facilitates [unit] testing by providing a valid, deterministic example of an AmbientWeatherDevice.

It's necessary because instantiating an AmbientWeatherDevice can [otherwise] only be done via Decoding, which requires the parent platform be surreptitiously passed in via a coding key that's internal to the library.  Rather than making that public - and requiring callers to have inappropriate knowledge of how these classes work internally - it's less evil to just provide this convenience constructor.