Removed the 'unit' property from WeatherSensor.
The unit is now solely conveyed by the measurement [value], which is either a Measurement with attached Unit or is a unitless value such as a Date or Int.

This also meant moving construction of the Measurement from the AmbientWeatherSensor initialiser(s) to AmbientWeatherReport.  This turned out to be a real pain in the arse - NSMeasurement is bridged to Swift *twice*, once as "Measurement" the _class_ and once as "Measurement" the _struct_.  There's no way to explicitly specify that you want to use the struct version - a major flaw in the type system, it seems.  In lieu of a clear selection, which one is chosen by the compiler is determined very indirectly based on how the measurements are used, which is a very complicated determination that the compiler frequently struggles with (just hanging entirely, unable to figure out which one to use) and that's very difficult to predict the outcome of.  Sigh.

I feel like using the struct version would be better in many ways - more strongly typed, more efficient perhaps, etc.  But it's just impossible as far as I can tell.  Or at least wildly impractical - I'd have to go to tremendous, absurd lengths involving type erasure (or viral generification of the entire library) to have even the slightest chance of getting it to compile.  Let-alone be usable in practice (e.g. it appears to be *impossible* to recover a struct Measurement from Any in some cases - and it's hard to even predict which cases that will be).

This is a "SwiftUI moment" for Swift, that I've hit quite often in the past.  Swift is great when you're doing relatively simple things, but then you try to do something just _slightly_ off the beaten path and it explodes spectacularly.

At times like these I _particularly_ miss Objective-C.