Used an "AnyMD" type eraser to allow the use of the actual Measurement struct, rather than class.
It's a huge pain in the arse to do this because of Swift's terrible ergonomics regarding generics, but it's [hopefully] worth it because it has one, significant benefit:  users can retrieve _either_ the struct or class forms of Measurement from the Any that they're provided via WeatherSensor!  Apparently Swift will magically do the conversion from struct to class form implicitly if you do a "value as? Measurement".  There's no implicit conversion _from_ class to struct form, though, which is why it's so important that the real underlying values be in struct form.  Plus it's presumably more efficient at runtime.