Slightly improved the mess around 'MD' and all the decoding stuff in AmbientWeatherReport, by moving the input value type into a generic parameter.
This allows for type information to be carried through further, which allows some code to be omitted and rules out certain kinds of internal inconsistencies (by making them compile-time errors instead of runtime).

This does mean the generic arguments basically need to be provided at every initialisation of an 'MD', but that's actually not bad - it's actually still less text than how it was, since ".self" isn't needed on the end of every type name and the units - for standard Unit subtypes, at least - simply shift from the init argument to the generic argument.

Best of all, as a consequence of _that_, no more "as AnyMD" suffixed on every initialisation of an MD [in the main list]!  Woot!