Finally fixed AmbientWeatherError's localised descriptions.
Previously the intended value - the human-readable, localisable string - was being used as the key, even though it was both (a) super verbose in many cases and (b) non-deterministic because it used string interpolation to include context / parameters.  Now the NSLocalizedString(…) function is _actually_ used correctly, with a plain, simple string literal as the key (matching the error case name) and the default (English) value provided correctly as the "value" argument.  Dynamic inclusion of arguments (context) is now also performed correctly via a wrapping String.localizedStringWithFormat(…) call.

Additionally, as part of cleaning all this up I rejiggered some of the error cases - and added some new ones - so that they would be clearer and writing meaningful, human-readable descriptions of them is easier.