Customised the URLSession configuration for AmbientWeather, and added a parameter to SwiftWeather.create(…) for overriding that.
The default configuration now:

• Uses HTTP pipelining [for increased efficiency].
• Waits for network connectivity if necessary.
• Much more sensible about timeouts (i.e. two minutes instead of seven days).
• Allows use of the network in all situations (may or may not be ideal for a specific app, but is a sensible default given how little bandwidth weather reports use).
• Doesn't allow the use of cookies.
• Does allow the use of cache but only if it the cached value is confirmed still good by the server.  I _assume_ this is using If-Modified-Since, although Apple's documentation is completely opaque about this.  I think this is the safest _default_ since its possible framework users will make the same requests repeatedly, though they might wish to override this if they know better.
• Requires TLS 1.3 at a minimum.
• Allows only one connection at a time [per host].  This will help avoid rate-limiting errors, though it is conceivable that it will be a bottleneck in some situations, if responses from AmbientWeather take more than a second or two.  We'll have to see how it pans out in practice.

The last two are enforced for _all_ configs, even if callers provide their own, as I think they're essential to secure, reliable operation.

Also removed the 'weatherServiceType' label from the first parameter to SwiftWeather.create(…), since it's very clear already, at the call site, what the argument is given it's an enum value.

Oh, and also, this meant maintaining Codable conformance for AmbientWeather was untenable, as there's no support for NSCoding or Codable on URLSessionConfiguration (or URLSession itself).  So it has been removed.  This shouldn't be a practical issue - URLSessionConfiguration is probably something you want to hard-code into the app itself, not its config, and the WeatherPlatformType enum is still Codable so there's still a good way to persist the essential configuration of a platform (including its application & API keys).