Removed Codable as a requirement for WeatherDevice.
As I work to add Wunderground support it's confirming that this is just more trouble than it's worth.

More importantly, perhaps, I'm not sure it makes sense to allow persistence of the device objects themselves.  My original thinking was that it could be convenient for using the API, especially with things like SwiftUI where you could just stuff it into user preferences.  And maybe that would be.  But the problem is that WeatherDevices are also expected to link back to their parent platform, so that'd have to be encoded, and I'm not sure that's a good road to go down.  Plus they might contain state which needs refreshing - i.e. potentially _everything_ other than their essential ID - and I don't want to commit them to having to do that at decode time (just think of the potential problems with random GUI hangs as a result of network calls in init(from: Decoder)!).

There's still some big gaps that need resolving in the absence of this - e.g. how _do_ you persist a WeatherDevice, functionally?  If it's an AmbientWeatherDevice you can only get that from the 'devices' property on the platform, but on Wunderground that method won't work and you'll have to specify it _manually_.  I don't like the the idea of having wildly different methodologies across weather platforms - it undermines having a unified interface.