Changed the 'devices' property 'usersDevices', on WeatherPlatform, to better reflect its purpose.
With AmbientWeather it's the _only_ way to obtain WeatherDevices, and it only covers those owned by the user (based on the API key in use).  Not all platforms have a notion of the current user's device(s) - or at least don't have APIs to retrieve them, such as Wunderground.  It's also impractical to try returning _all_ the devices on most platforms since there are countless of them and the vast majority surely aren't of interest to the user.

There'll have to be different, additional APIs for finding devices on other platforms (e.g. based on a location, or textual search term).