Made WeatherDevice generic by adding a Platform associated type.
This allows its 'platform' property to return a concrete type, rather than a type-erased WeatherPlatform.  Basically I'm dabbling with the idea of making all the protocols generic, down to the sensors and their associated unit types, as a way to make it easier to work with _specific_ platforms.  Currently the various layers of type erasure make things very "interchangeable" but a bit awkward to actually use in other respects.

This will require the use of `any WeatherDevice` in client code, if they actually want that type erasure.  It's a little annoying perhaps, but arguably they should have been using that anyway to make it clear that type erasure is occurring.