Work-in-progress implementation of Wunderground support.
It currently supports fetching stations near a geolocation, and fetching the latest report for a station.

Unfortunately the Wunderground API is a bit over-designed and difficult to use, and doesn't provide some fundamental capabilities such as retrieving historical reports.  It does have a way to retrieve historical _summaries_ - actually about three or four ways, confusingly, each with slight and annoying differences in behaviour - but their documentation is incomplete and I haven't tried them yet, so I'm not entirely sure how they work.  I expect it'll be significant work to add support for those, which is why I'm checking this in now [so that I don't have to keep working around it in my Git commits].

There's also an API - currently #if false'd out by default - for fetching nearby weather devices.  It relies on Core Location.  The idea is that it's a convenience method for the very common use-case of just wanting to find the nearest weather station(s).  Unfortunately Core Location appears to be an incredibly buggy piece of shit, and despite days of screwing around with it and applying hacky workaround after hacky workaround, I'm still unable to get it to work reliably in any way.  I tested several other random projects, from GitHub, that use it as well, and they all don't work either.  So this might be something about my specific computer.  But location services work in every actual app I've ever used otherwise, like Maps as well as 3rd party weather apps, so I'm wondering if it's some bullshit problem with code signing or entitlements or… who knows.  Sigh.