Made some of the custom units adhere to their actual Dimension types.
Namely, microgramsPerCubicMetre is now properly a UnitConcentrationMass, and both millimetresPerHour & inchesPerHour are now properly UnitSpeeds.

This does have unfortunate consequences for value formatting, though - UnitSpeed defaults to things like "m/s" or "km/h" which are wildly inappropriate scales for rainfall.  But this isn't the first case of measurements being improperly presented - e.g. all the cumulative rain numbers are presented as metres by default, which is also verging on useless (especially if there are few decimal digits to provide the necessary resolution).  I think the solution is going to have to be a lot of custom formatting hints etc in WeatherSensor's FormatStyle.  I explored this briefly just now, but I'm undecided as to which way to go (given the whole NSMeasurementFormatter vs Measurement<Dimension>.formatted(…) etc).  I left a TODO to revisit this later.